+++ Cargo.toml
@@ -73,7 +73,7 @@ web-time = "1.1"
 # - https://github.com/hyperium/tonic/pull/1974
 # - https://github.com/rustls/rustls-platform-verifier/issues/58
 async-compression = { default-features = false, version = "0.4", features = [
-  "tokio",
+  "futures-io",
   "zstd",
 ] }
 bincode = "1.3"
@@ -113,6 +113,7 @@ xmtp_cryptography = { path = "xmtp_cryptography" }
 xmtp_id = { path = "xmtp_id" }
 xmtp_mls = { path = "xmtp_mls" }
 xmtp_proto = { path = "xmtp_proto" }
+
 
 [profile.dev]
 # Disabling debug info speeds up builds a bunch,
