+++ xmtp_mls/src/groups/device_sync/backup.rs
@@ -54,6 +54,7 @@ impl BackupOptions {
 }
 
 #[cfg(test)]
+#[cfg(not(target_arch = "wasm32"))]
 mod tests {
     use super::*;
     use crate::{builder::ClientBuilder, groups::GroupMetadataOptions};
