+++ xmtp_mls/src/groups/device_sync/backup/backup_importer.rs
@@ -1,3 +1,4 @@
+use super::BackupError;
 use crate::{
     groups::device_sync::DeviceSyncError,
     storage::{
@@ -6,13 +7,15 @@ use crate::{
     },
     Store, XmtpOpenMlsProvider,
 };
-use async_compression::tokio::bufread::ZstdDecoder;
+use async_compression::futures::bufread::ZstdDecoder;
+use futures::{io::BufReader, AsyncBufRead, AsyncRead, AsyncReadExt};
 use prost::Message;
 use std::pin::Pin;
-use tokio::io::{AsyncBufRead, AsyncRead, AsyncReadExt, BufReader};
+use tokio_util::compat::{FuturesAsyncReadCompatExt, TokioAsyncReadCompatExt};
 use xmtp_proto::xmtp::device_sync::{backup_element::Element, BackupElement, BackupMetadata};
 
-use super::BackupError;
+#[cfg(not(target_arch = "wasm32"))]
+mod file_import;
 
 pub struct BackupImporter {
     decoded: Vec<u8>,
@@ -21,28 +24,6 @@ pub struct BackupImporter {
 }
 
 impl BackupImporter {
-    pub async fn open(reader: impl AsyncRead + Send + 'static) -> Result<Self, DeviceSyncError> {
-        let reader = BufReader::new(reader);
-        let reader = Box::pin(reader) as Pin<Box<_>>;
-        let decoder = ZstdDecoder::new(reader);
-
-        let mut importer = Self {
-            decoder,
-            decoded: vec![],
-            metadata: BackupMetadata::default(),
-        };
-
-        let Some(BackupElement {
-            element: Some(Element::Metadata(metadata)),
-        }) = importer.next_element().await?
-        else {
-            return Err(BackupError::MissingMetadata)?;
-        };
-
-        importer.metadata = metadata;
-        Ok(importer)
-    }
-
     async fn next_element(&mut self) -> Result<Option<BackupElement>, StorageError> {
         let mut buffer = [0u8; 1024];
         let mut element_len = 0;
